pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pcb-defect-detection"
version = "0.2.0"
description = "End-to-end tooling for training, evaluating, and deploying PCB defect detectors"
authors = [{ name = "Siemon Cha", email = "maintainers@pcb-defect-detection.example" }]
maintainers = [{ name = "PCB Defect Detection Maintainers", email = "maintainers@pcb-defect-detection.example" }]
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
keywords = ["pcb", "computer-vision", "defect-detection", "ultralytics", "fastapi"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Manufacturing",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
  "fastapi",
  "matplotlib",
  "numpy",
  "onnxruntime",
  "onnxslim",
  "opencv-python",
  "pillow>=9.0",
  "pyyaml",
  "requests",
  "scipy",
  "seaborn",
  "typer",
  "ultralytics>=8.0.0",
  "uvicorn[standard]",
]

[project.optional-dependencies]
analytics = [
  "pandas",
  "polars",
]
dev = [
  "pytest",
  "pytest-mock",
  "ruff",
]

[project.scripts]
pcb-dd = "pcb_defect_detection.cli:main"

[project.urls]
"Documentation" = "https://github.com/siemoncha/pcb-defect-detection"
"Issue Tracker" = "https://github.com/siemoncha/pcb-defect-detection/issues"
"Repository" = "https://github.com/siemoncha/pcb-defect-detection"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"pcb_defect_detection.config" = ["default.yaml"]